requirements.txt
# pip install -r requirements.txt

numpy==1.19.5                             # last version supported - 20/oct/21
scipy==1.5.4                              # last version supported - 20/oct/21
scikit-learn==0.24.2                      # last version supported - 20/oct/21
microtc==2.2.9                            # last version supported - 20/oct/21
pandas==1.1.5                             # last version supported - 20/oct/21
apache-airflow[postgres]==2.1.4